+++ structure/bbst/reversible-splay-tree.hpp
@@ -1,180 +1,51 @@
 /**
  * @brief Reversible-Splay-Tree(反転可能Splay木)
  */
-template <typename Monoid = int, typename OperatorMonoid = void>
-struct ReversibleSplayTree {
+template <typename Tp>
+struct ReversibleSplayTreeNode {
+  using T = Tp;
+  ReversibleSplayTreeNode *l, *r, *p;
+  T key, sum;
+  bool rev;
+  size_t sz;
+
+  ReversibleSplayTreeNode() : ReversibleSplayTreeNode(Tp()) {}
+
+  ReversibleSplayTreeNode(const T &key)
+      : key(key),
+        sum(key),
+        rev(false),
+        l(nullptr),
+        r(nullptr),
+        p(nullptr),
+        sz(1) {}
+};
+
+template <typename Np>
+struct ReversibleSplayTree : SplayTreeBase<Np> {
  public:
-  using F = function<Monoid(Monoid, Monoid)>;
-  using S = function<Monoid(Monoid)>;
+  using Node = Np;
+  using T = typename Node::T;
+  using F = function<T(T, T)>;
+  using S = function<T(T)>;
+  using super = SplayTreeBase<Node>;
+  using NP = typename super::NP;
 
-  struct Node {
-    Node *l, *r, *p;
-    Monoid key, sum;
-    bool rev;
-    size_t sz;
-
-    bool is_root() const { return !p || (p->l != this && p->r != this); }
-
-    Node(const Monoid &key)
-        : key(key),
-          sum(key),
-          sz(1),
-          rev(false),
-          l(nullptr),
-          r(nullptr),
-          p(nullptr) {}
-  };
-
-  ReversibleSplayTree(const F &f, const Monoid &M1)
-      : ReversibleSplayTree(f, [](const Monoid &a) { return a; }, M1) {}
-
-  ReversibleSplayTree(const F &f, const S &s, const Monoid &M1)
+  explicit ReversibleSplayTree(const F &f, const S &s, const T &M1)
       : f(f), s(s), M1(M1) {}
 
-  inline size_t count(const Node *t) { return t ? t->sz : 0; }
+  using super::build_node;
+  using super::count;
+  using super::insert_node;
+  using super::merge;
+  using super::splay;
+  using super::split;
 
-  inline const Monoid &sum(const Node *t) { return t ? t->sum : M1; }
+  inline const T &sum(const NP t) { return t ? t->sum : M1; }
 
-  Node *alloc(const Monoid &v = Monoid()) { return new Node(v); }
+  NP alloc(const T &x) { return new Node(x); }
 
-  void splay(Node *t) {
-    push(t);
-    while (!t->is_root()) {
-      auto *q = t->p;
-      if (q->is_root()) {
-        push(q), push(t);
-        if (q->l == t)
-          rotr(t);
-        else
-          rotl(t);
-      } else {
-        auto *r = q->p;
-        push(r), push(q), push(t);
-        if (r->l == q) {
-          if (q->l == t)
-            rotr(q), rotr(t);
-          else
-            rotl(t), rotr(t);
-        } else {
-          if (q->r == t)
-            rotl(q), rotl(t);
-          else
-            rotr(t), rotl(t);
-        }
-      }
-    }
-  }
-
-  Node *push_front(Node *t, const Monoid &v = Monoid()) {
-    if (!t) {
-      t = alloc(v);
-      return t;
-    } else {
-      splay(t);
-      Node *cur = get_left(t), *z = alloc(v);
-      splay(cur);
-      z->p = cur;
-      cur->l = z;
-      splay(z);
-      return z;
-    }
-  }
-
-  Node *push_back(Node *t, const Monoid &v = Monoid()) {
-    if (!t) {
-      t = alloc(v);
-      return t;
-    } else {
-      splay(t);
-      Node *cur = get_right(t), *z = alloc(v);
-      splay(cur);
-      z->p = cur;
-      cur->r = z;
-      splay(z);
-      return z;
-    }
-  }
-
-  Node *erase(Node *t) {
-    splay(t);
-    Node *x = t->l, *y = t->r;
-    delete t;
-    if (!x) {
-      t = y;
-      if (t) t->p = nullptr;
-    } else if (!y) {
-      t = x;
-      t->p = nullptr;
-    } else {
-      x->p = nullptr;
-      t = get_right(x);
-      splay(t);
-      t->r = y;
-      y->p = t;
-    }
-    return t;
-  }
-
-  Node *get_left(Node *t) const {
-    while (t->l) t = t->l;
-    return t;
-  }
-
-  Node *get_right(Node *t) const {
-    while (t->r) t = t->r;
-    return t;
-  }
-
-  pair<Node *, Node *> split(Node *t, int k) {
-    if (!t) return {nullptr, nullptr};
-    push(t);
-    if (k <= count(t->l)) {
-      auto x = split(t->l, k);
-      t->l = x.second;
-      t->p = nullptr;
-      if (x.second) x.second->p = t;
-      return {x.first, update(t)};
-    } else {
-      auto x = split(t->r, k - count(t->l) - 1);
-      t->r = x.first;
-      t->p = nullptr;
-      if (x.first) x.first->p = t;
-      return {update(t), x.second};
-    }
-  }
-
-  template <typename... Args>
-  Node *merge(Node *l, Args... rest) {
-    Node *r = merge(rest...);
-    if (!l && !r) return nullptr;
-    if (!l) return splay(r), r;
-    if (!r) return splay(l), l;
-    splay(l), splay(r);
-    l = get_right(l);
-    splay(l);
-    l->r = r;
-    r->p = l;
-    update(l);
-    return l;
-  }
-
-  void insert(Node *&t, int k, const Monoid &v) {
-    splay(t);
-    auto x = split(t, k);
-    t = merge(merge(x.first, alloc(v)), x.second);
-  }
-
-  Monoid erase(Node *&t, int k) {
-    splay(t);
-    auto x = split(t, k);
-    auto y = split(x.second, 1);
-    auto v = y.first->c;
-    delete y.first;
-    t = merge(x.first, y.second);
-    return v;
-  }
-
-  Monoid query(Node *&t, int a, int b) {
+  T query(NP &t, int a, int b) {
     splay(t);
     auto x = split(t, a);
     auto y = split(x.second, b - a);
@@ -183,15 +54,19 @@ struct ReversibleSplayTree {
     return ret;
   }
 
-  Node *build(const vector<Monoid> &v) { return build(0, (int)v.size(), v); }
+  NP build(const vector<T> &v) {
+    vector<NP> vs(v.size());
+    for (int i = 0; i < v.size(); i++) vs[i] = alloc(v[i]);
+    return build_node(vs);
+  }
 
-  void toggle(Node *t) {
+  void toggle(NP t) {
     swap(t->l, t->r);
     t->sum = s(t->sum);
     t->rev ^= true;
   }
 
-  Node *update(Node *t) {
+  NP update(NP t) override {
     t->sz = 1;
     t->sum = t->key;
     if (t->l) t->sz += t->l->sz, t->sum = f(t->l->sum, t->sum);
@@ -199,14 +74,7 @@ struct ReversibleSplayTree {
     return t;
   }
 
-  tuple<Node *, Node *, Node *> split3(Node *t, int a, int b) {
-    splay(t);
-    auto x = split(t, a);
-    auto y = split(x.second, b - a);
-    return make_tuple(x.first, y.first, y.second);
-  }
-
-  void push(Node *t) {
+  void push(NP t) override {
     if (t->rev) {
       if (t->l) toggle(t->l);
       if (t->r) toggle(t->r);
@@ -214,57 +82,49 @@ struct ReversibleSplayTree {
     }
   }
 
-  void set_element(Node *&t, int k, const Monoid &x) {
+  NP insert(NP t, int k, const T &x) { return insert_node(t, k, alloc(x)); }
+
+  NP set_element(NP t, int k, const T &x) {
     splay(t);
-    sub_set_element(t, k, x);
+    return imp_set_element(t, k, x);
+  }
+
+  pair<NP, NP> split_lower_bound(NP t, const T &key) {
+    if (!t) return {nullptr, nullptr};
+    push(t);
+    if (key <= t->key) {
+      auto x = split_lower_bound(t->l, key);
+      t->l = x.second;
+      t->p = nullptr;
+      if (x.second) x.second->p = t;
+      return {x.first, update(t)};
+    } else {
+      auto x = split_lower_bound(t->r, key);
+      t->r = x.first;
+      t->p = nullptr;
+      if (x.first) x.first->p = t;
+      return {update(t), x.second};
+    }
   }
 
  private:
-  const Monoid M1;
+  const T M1;
   const F f;
   const S s;
 
-  Node *build(int l, int r, const vector<Monoid> &v) {
-    if (l + 1 >= r) return alloc(v[l]);
-    return merge(build(l, (l + r) >> 1, v), build((l + r) >> 1, r, v));
-  }
-
-  Node *sub_set_element(Node *&t, int k, const Monoid &x) {
+  NP imp_set_element(NP t, int k, const T &x) {
     push(t);
     if (k < count(t->l)) {
-      return sub_set_element(t->l, k, x);
+      return imp_set_element(t->l, k, x);
     } else if (k == count(t->l)) {
       t->key = x;
       splay(t);
       return t;
     } else {
-      return sub_set_element(t->r, k - count(t->l) - 1, x);
+      return imp_set_element(t->r, k - count(t->l) - 1, x);
     }
   }
-
-  void rotr(Node *t) {
-    auto *x = t->p, *y = x->p;
-    if ((x->l = t->r)) t->r->p = x;
-    t->r = x, x->p = t;
-    update(x), update(t);
-    if ((t->p = y)) {
-      if (y->l == x) y->l = t;
-      if (y->r == x) y->r = t;
-      update(y);
-    }
-  }
-
-  void rotl(Node *t) {
-    auto *x = t->p, *y = x->p;
-    if ((x->r = t->l)) t->l->p = x;
-    t->l = x, x->p = t;
-    update(x), update(t);
-    if ((t->p = y)) {
-      if (y->l == x) y->l = t;
-      if (y->r == x) y->r = t;
-      update(y);
-    }
-  }
-
-  Node *merge(Node *l) { return l; }
 };
+
+template <typename T>
+using RST = ReversibleSplayTree<ReversibleSplayTreeNode<T> >;
